# ZIO Test: Improve Performance of ZIO Internal Test Suite on ScalaJS