--- a/test/js/src/test/scala/zio/test/TestPlatform.scala
+++ b/test/js/src/test/scala/zio/test/TestPlatform.scala
@@ -2,10 +2,14 @@ package zio.test
 
 import scala.concurrent.{ ExecutionContext, Future }
 
-object ExitUtils {
-
-  def fail(): Unit = System.exit(-1)
+private[test] object TestPlatform {
 
   def await(f: Future[Boolean])(implicit ec: ExecutionContext): Unit =
     f.foreach(_ => ())
+
+  def fail(): Unit = System.exit(-1)
+
+  val isJS = true
+
+  val isJVM = false
 }