--- a/test/jvm/src/test/scala/zio/test/ExitUtils.scala
+++ b/test/jvm/src/test/scala/zio/test/ExitUtils.scala
@@ -3,16 +3,12 @@ package zio.test
 import scala.concurrent.{ Await, ExecutionContext, Future }
 import scala.concurrent.duration._
 
-private[test] object TestPlatform {
+private[test] object ExitUtils {
 
-  def await(f: Future[Boolean])(implicit ec: ExecutionContext): Unit = {
+  final def await(f: Future[Boolean])(implicit ec: ExecutionContext): Unit = {
     val passed = Await.result(f.map(identity), 60.seconds)
     if (passed) () else throw new AssertionError("tests failed")
   }
 
-  def fail(): Unit = ()
-
-  val isJS = false
-
-  val isJVM = true
+  final def fail(): Unit = ()
 }