--- a/test/jvm/src/test/scala/zio/test/TestPlatform.scala
+++ b/test/jvm/src/test/scala/zio/test/TestPlatform.scala
@@ -3,12 +3,16 @@ package zio.test
 import scala.concurrent.{ Await, ExecutionContext, Future }
 import scala.concurrent.duration._
 
-object ExitUtils {
-
-  def fail(): Unit = ()
+private[test] object TestPlatform {
 
   def await(f: Future[Boolean])(implicit ec: ExecutionContext): Unit = {
     val passed = Await.result(f.map(identity), 60.seconds)
     if (passed) () else throw new AssertionError("tests failed")
   }
+
+  def fail(): Unit = ()
+  
+  val isJS = false
+
+  val isJVM = true
 }