--- a/test/jvm/src/test/scala/zio/test/TestPlatform.scala
+++ b/test/jvm/src/test/scala/zio/test/TestPlatform.scala
@@ -11,7 +11,7 @@ private[test] object TestPlatform {
   }
 
   def fail(): Unit = ()
-  
+
   val isJS = false
 
   val isJVM = true