--- a/test/shared/src/test/scala/zio/test/TestMain.scala
+++ b/test/shared/src/test/scala/zio/test/TestMain.scala
@@ -22,7 +22,7 @@ object TestMain {
       scope(SampleSpec.run, "SampleSpec"),
       scope(SchedulerSpec.run, "SchedulerSpec"),
       scope(SystemSpec.run, "SystemSpec")
-    ).flatten
+    )
     report(testResults)
   }
 }