--- a/test/shared/src/test/scala/zio/test/TestMain.scala
+++ b/test/shared/src/test/scala/zio/test/TestMain.scala
@@ -21,7 +21,8 @@ object TestMain {
       scope(RandomSpec.run, "RandomSpec"),
       scope(SampleSpec.run, "SampleSpec"),
       scope(SchedulerSpec.run, "SchedulerSpec"),
-      scope(SystemSpec.run, "SystemSpec")
+      scope(SystemSpec.run, "SystemSpec"),
+      scope(TestAspectSpec.run, "TestAspectSpec")
     )
     report(testResults)
   }