--- a/test/shared/src/test/scala/zio/test/TestUtils.scala
+++ b/test/shared/src/test/scala/zio/test/TestUtils.scala
@@ -7,39 +7,42 @@ import zio.clock.Clock
 
 object TestUtils {
 
-  def label(f: Future[Boolean], s: String)(implicit ec: ExecutionContext): Future[(Boolean, String)] =
-    f.map { p =>
-      if (p)
-        (p, succeed(s))
-      else
-        (p, fail(s))
-    }.recover { case _ => (false, fail(s)) }
+  final def label(test: => Future[Boolean], label: String): Async[(Boolean, String)] =
+    Async
+      .fromFuture(test)
+      .map(passed => if (passed) (passed, succeed(label)) else (passed, fail(label)))
+      .handle { case _ => (false, fail(label)) }
 
-  def nonFlaky[R, E](test: ZIO[R, E, Boolean]): ZIO[R with Clock, E, Boolean] =
+  final def nonFlaky[R, E](test: ZIO[R, E, Boolean]): ZIO[R with Clock, E, Boolean] =
     test.repeat(Schedule.recurs(100) *> Schedule.identity[Boolean])
 
-  def scope(fs: List[Future[(Boolean, String)]], s: String)(
-    implicit ec: ExecutionContext
-  ): List[Future[(Boolean, String)]] = {
-    val p      = Future.sequence(fs).map(_.forall(_._1))
-    val offset = fs.map(_.map { case (p, s) => (p, "  " + s) })
-    p.map(p => if (p) (p, succeed(s)) else (p, fail(s))) :: offset
+  final def report(suites: List[Async[List[(Boolean, String)]]])(implicit ec: ExecutionContext): Unit = {
+    val async = Async
+      .sequence(suites)
+      .map(_.flatten)
+      .flatMap { results =>
+        val passed = results.forall(_._1)
+        results.foreach(result => println(result._2))
+        if (passed) Async.succeed(true)
+        else Async(ExitUtils.fail()).map(_ => false)
+      }
+    ExitUtils.await(async.run(ec))
   }
 
-  def report(ps: List[Future[(Boolean, String)]])(implicit ec: ExecutionContext): Unit = {
-    val f = Future
-      .sequence(ps)
-      .map(results => (results.forall(_._1), results))
-      .flatMap {
-        case (passed, results) =>
-          results.foreach(result => println(result._2))
-          if (!passed)
-            Future(ExitUtils.fail()).map(_ => false)
-          else
-            Future.successful(true)
-      }
-    ExitUtils.await(f)
-  }
+  final def scope(tests: List[Async[(Boolean, String)]], label: String): Async[List[(Boolean, String)]] =
+    Async.sequence(tests).map { tests =>
+      val offset = tests.map { case (passed, label) => (passed, "  " + label) }
+      val passed = tests.forall(_._1)
+      if (passed) (passed, succeed(label)) :: offset else (passed, fail(label)) :: offset
+    }
+
+  final def timeit[A](label: String)(async: Async[A]): Async[A] =
+    for {
+      start  <- Async(System.currentTimeMillis)
+      result <- async
+      stop   <- Async(System.currentTimeMillis)
+      _      <- Async(println(s"$label took ${(stop - start) / 1000.0} seconds"))
+    } yield result
 
   private def succeed(s: String): String =
     green("+") + " " + s