We have been experiencing slow builds on ScalaJS due to the ZIO Test test suite. On the JVM the full test suite runs in about 30 seconds, we have a timeout of 60 seconds for the entire test suite, and to my knowledge this has never caused flakiness. In contrast, on ScalaJS we are observing running the test suite taking 4 to 5 minutes.

It appears that there are several root causes for this:

**1. Possible Excessive Parallelism** - We use `Future` as our asynchronous effect type for tests since we don't want to rely on ZIO while testing ZIO and have been using `sequence` to combine all the tests together. `Future.sequence` runs effects in parallel, so we may have had a large number of tests running in parallel, causing contention that was handled poorly in the ScalaJS single threaded environment. This PR addresses that by adding an `Async` type that provides a thin, purely functional wrapper around `Future` so that we can properly sequence effects. This appears to help in my local tests though I am interested to see what impact it has on CI build times.

**2. Poor Performance of Random Number Generation in ScalaJS** - It appears that random number generation in Scala.js is substantially slower than on the JVM. In benchmarking on my local machine random number generation takes approximately three times as long on ScalaJS as on the JVM. Note that this has nothing to do with `MockRandom`, this is just performance of `scala.util.Random` from the standard library and Scala.js. The parts of the test suite that take the longest all rely heavily on random number generation, particularly `GenSpec`, so I think this will explain most of the performance difference after the parallelism issue is fixed.

**3. Potential ZIO Performance Differences Across Platforms** - There may be differences in ZIO performance across platforms. For example, the snippet below ran in about 4 seconds for me on the JVM but took 36 seconds on ScalaJS. The random number generation issue is definitely a part of that but there may also be other cross platform differences. I haven't seen any comparison of ZIO performance across platforms so I don't have a sense of what we should expect here and how things like whether you do full optimization in ScalaJS impact that.

**4. Some Tests Take Disproportionate Time** - The tests that take the longest verify that fixed size collections shrink appropriately (all elements of the collection shrink to their "smallest" value). We currently do this on size 10 lists, vectors, and strings and repeat each test 100 times to verify non-flakiness. We could definitely cut this down though I do think these are important tests and reducing the number of tests is to a certain extent a one time improvement that doesn't address the root causes, especially since the time for the entire test suite is very reasonable on the JVM.

Let's see how much we get out of these changes and then we can make further adjustments if necessary.

```scala
def benchmark = 
  unsafeRunToFuture {
    import zio._
    val io = ZIO.collectAll(List.fill(100000)(random.nextInt(100))).repeat(Schedule.recurs(100))
    io.timed.tap { case (duration, _) => ZIO.effectTotal(println(duration.toMillis / 1000.0)) }
  }
```