@ghostdogpr 2019-08-31
Looking at the test runs, it seems the first change (Async) was not that impactful (still around 4min, sometimes 5min), but the second one (nonflakiness) improved it a lot (~1min30).