@adamgfraser 2019-08-31
That was my conclusion as well. I am going to try to get the `main` method of `TestMain` to accept a command line `size` parameter so we can pass a different `size` in for `testJVM` and `testJS`. Then the test suites can use that parameter how they want, so we can still run the additional tests on the JVM. I think that will strike a nice balance between getting the times down and maintaining coverage. And if contributors are making changes and testing interactively in the `testJVM` project they will still get strong assurances that everything is working correctly.