@adamgfraser 2019-09-01
Okay, well my plan to use a command line argument kind of blew up in my face when I learned that ScalaJS doesn't accept command line arguments but I developed a workaround. This is ready to go from my perspective.