@jdegoes 2019-09-01
So the reason why `Async` cannot help is because (among other things), Javascript only has 1 thread, so `sequence` cannot introduce "too much concurrency". Even being lazy about when these futures are started (construction versus "unsafeRun") is not going to impact performance. What will help is fewer futures: i.e. doing more inside each future, so you need fewer flat maps. It would also help to use ZIO instead of Future, although that's not possible / viable due to bootstrapping.

All monadic-heavy code will be quite a bit slower on Javascript. A bit surprising how much, though, and possibly worth profiling and doing a deep dive, but not now, I'd say (too little ROI).

I think we could benefit from `isJs` and `isJVM` being inside ZIO Test directly, btw.