@adamgfraser 2019-09-01
@jdegoes I incorporated your feedback, including the test aspects, here. Resolves #1541.