@ghostdogpr 2019-09-02
@adamgfraser I fixed a typo to re-trigger the build, the CI is back. I'll merge if the tests pass.