@ghostdogpr 2019-09-02
There's some issue downloading JDK from CircleCI, will re-try again later :/